requirements.txt
openai>=1.55.3
pillow
pytesseract
python-docx
python-pptx
openpyxl
xlrd
